refactor: moved ErrorMessage to effect
