+++ src/message.js
@@ -19,11 +19,6 @@ export type Message = {
   +tag: MessageTag,
 };
 
-export type ErrorMessage = {
-  +tag: typeof EFFECT_ERROR,
-  +error: mixed,
-};
-
 /**
  * A message on its way upwards in the hierarchy.
  */
@@ -70,8 +65,6 @@ export type Subscription<M: Message> = true | {
    */
   matching?: MessageFilter<M>,
 };
-
-export const EFFECT_ERROR: "effect/error" = "effect/error";
 
 /**
  * @param {!Object} subscriptions
